+++ apps/CASTool_USEPA/global.R
@@ -5,7 +5,7 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~
 
 # Version ----
-pkg_version <- "1.0.0.9018"
+pkg_version <- "1.0.0.9019"
 
 # Packages ----
 library(shiny)